reduce dependency and code duplication between join and GC aligner
